Separate TokenBarsOverlayFeature into a separate feature from the token tracker bell
